refactor(geom): migrate PathTransformation hierarchy to DoubleArray
Convert all transformation implementations to use Double parameters and DoubleArray matrices.
Remove the Float-to-Double workaround in ArcTransformation.decomposeEllipseMatrix.

+++ svg-to-compose/src/commonMain/kotlin/dev/tonholo/s2c/geom/transform/LineTransformation.kt
@@ -5,8 +5,8 @@ import dev.tonholo.s2c.geom.AffineTransformation
 
 internal data object LineTransformation : PathTransformation<PathNodes.LineTo>() {
     override fun PathNodes.LineTo.applyTransformation(
-        cursor: FloatArray,
-        start: FloatArray,
+        cursor: DoubleArray,
+        start: DoubleArray,
         transformation: AffineTransformation,
     ): PathNodes {
         val (x, y) = if (isRelative) {
